強制更新モード追加

+++ PsUpdate.psm1
@@ -1,8 +1,8 @@
 ﻿#################################################################
 # PowerShell 更新
 #################################################################
-Function PsUpdate(){
-	if(Get-Command winget -ea SilentlyContinue){
+Function PsUpdate([switch]$Force){
+	if((Get-Command winget -ea SilentlyContinue) -and (-Not $Force)){
 		winget install --id Microsoft.Powershell --source winget
 	}
 	else{
